Remove polyfuse from the GND output and increase value of the 3.3V polyfuse
The polyfuse on the GND header interferes with things too much. The added
voltage drop affects measured analogue values and also limits the maximum
current available to source from all of the outputs as the current has to
sink into the GND connector.

The 3.3V polyfuse was too small and like the GND one would have limited
the maximum current which could be sinked into the outputs. I've increased
it massively and will let the SMPS current limit take care of any shorts.
The polyfuse is now there to protect against 12V being input into the 3.3V
header.

The GND header doesn't have any protection against having 12V injected into
it. It relies solely on some form of current limiting on the source of the
12V being injected (which all available 12V lines have).